First implementation that passed unit tests of fix for issue 463

+++ DSCResources/MSFT_xDSCWebService/PSWSIISEndpoint.psm1
@@ -1,6 +1,8 @@
 # This module file contains a utility to perform PSWS IIS Endpoint setup
 # Module exports New-PSWSEndpoint function to perform the endpoint setup
 
+New-Variable -Name DefaultAppPoolName  -Value 'PSWS' -Option ReadOnly -Force -Scope Script
+
 <#
     .SYNOPSIS
         Validate supplied configuration to setup the PSWS Endpoint Function
@@ -12,6 +14,10 @@ function Initialize-Endpoint
     [CmdletBinding()]
     param
     (
+        [Parameter()]
+        [System.String]
+        $appPool,
+
         [Parameter()]
         [System.String]
         $site,
@@ -101,13 +107,12 @@ function Initialize-Endpoint
 
     Test-IISInstall
 
-    $appPool = 'PSWS'
-
-    Write-Verbose -Message 'Delete the App Pool if it exists'
-    Remove-AppPool -apppool $appPool
-
-    Write-Verbose -Message 'Remove the site if it already exists'
+    # First remove the site so that the binding count on the application pool is reduced
+    Write-Verbose -Message "Remove the site [$site] if it already exists"
     Update-Site -siteName $site -siteAction Remove
+
+    Write-Verbose -Message "Delete App Pool [$appPool] if it exists"
+    Remove-AppPool -apppool $appPool
 
     # Check for existing binding, there should be no binding with the same port
     $allWebBindingsOnPort = Get-WebBinding | Where-Object -FilterScript {
@@ -127,9 +132,25 @@ function Initialize-Endpoint
         }
     }
 
-    Copy-PSWSConfigurationToIISEndpointFolder -path $path -cfgfile $cfgfile -svc $svc -mof $mof -dispatch $dispatch -asax $asax -dependentBinaries $dependentBinaries -language $language -dependentMUIFiles $dependentMUIFiles -psFiles $psFiles
+    Copy-PSWSConfigurationToIISEndpointFolder -path $path `
+        -cfgfile $cfgfile `
+        -svc $svc `
+        -mof $mof `
+        -dispatch $dispatch `
+        -asax $asax `
+        -dependentBinaries $dependentBinaries `
+        -language $language `
+        -dependentMUIFiles $dependentMUIFiles `
+        -psFiles $psFiles
 
-    New-IISWebSite -site $site -path $path -port $port -app $app -apppool $appPool -applicationPoolIdentityType $applicationPoolIdentityType -certificateThumbPrint $certificateThumbPrint -enable32BitAppOnWin64 $enable32BitAppOnWin64
+    New-IISWebSite -site $site `
+        -path $path
+        -port $port `
+        -app $app `
+        -apppool $appPool `
+        -applicationPoolIdentityType $applicationPoolIdentityType `
+        -certificateThumbPrint $certificateThumbPrint `
+        -enable32BitAppOnWin64 $enable32BitAppOnWin64
 }
 
 <#
@@ -232,6 +253,40 @@ function Update-Site
 
 <#
     .SYNOPSIS
+        Returns the list of bound sites and applications for a given IIS Application pool
+
+    .PARAMETER appPool
+        The application pool name
+#>
+function Get-AppPoolBinding
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.String]
+        $appPool
+    )
+
+    if (Test-Path -Path "IIS:\AppPools\$appPool")
+    {
+        $sites = Get-WebConfigurationProperty `
+            -Filter "/system.applicationHost/sites/site/application[@applicationPool=`'$appPool`'and @path='/']/parent::*" `
+            -PSPath 'machine/webroot/apphost' `
+            -Name name
+        $apps = Get-WebConfigurationProperty `
+            -Filter "/system.applicationHost/sites/site/application[@applicationPool=`'$appPool`'and @path!='/']" `
+            -PSPath 'machine/webroot/apphost' `
+            -Name path
+        $sites, $apps | ForEach-Object {
+            $_.Value
+        }
+    }
+}
+
+<#
+    .SYNOPSIS
         Delete the given IIS Application Pool. This is required to cleanup any
         existing conflicting apppools before setting up the endpoint.
 #>
@@ -240,15 +295,31 @@ function Remove-AppPool
     [CmdletBinding()]
     param
     (
-        [Parameter()]
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
         [System.String]
         $appPool
     )
 
-    # Without this tests we may get a breaking error here, despite SilentlyContinue
-    if (Test-Path -Path "IIS:\AppPools\$appPool")
+    if ($DefaultAppPoolName -eq $appPool)
     {
-        Remove-WebAppPool -Name $appPool -ErrorAction SilentlyContinue
+        # Without this tests we may get a breaking error here, despite SilentlyContinue
+        if (Test-Path -Path "IIS:\AppPools\$appPool")
+        {
+            $cntBindings = (Get-AppPoolBinding -apppool $appPool | Measure-Object).Count
+            if (0 -ge $cntBindings)
+            {
+                Remove-WebAppPool -Name $appPool -ErrorAction SilentlyContinue
+            }
+            else
+            {
+                Write-Verbose -Message "Unable to delete application pool [$appPool] because it's still bound to a site or application"
+            }
+        }
+    }
+    else
+    {
+        Write-Verbose -Message "ApplicationPool name is different from built in name [$DefaultAppPoolName]. Unable to delete."
     }
 }
 
@@ -406,33 +477,50 @@ function New-IISWebSite
 
     $siteID = New-SiteID
 
-    Write-Verbose -Message 'Adding App Pool'
-    $null = New-WebAppPool -Name $appPool
-
-    Write-Verbose -Message 'Set App Pool Properties'
-    $appPoolIdentity = 4
-    if ($applicationPoolIdentityType)
+    if (Test-Path IIS:\AppPools\$appPool)
     {
-        # LocalSystem = 0, LocalService = 1, NetworkService = 2, SpecificUser = 3, ApplicationPoolIdentity = 4
-        if ($applicationPoolIdentityType -eq 'LocalSystem')
-        {
-            $appPoolIdentity = 0
-        }
-        elseif ($applicationPoolIdentityType -eq 'LocalService')
-        {
-            $appPoolIdentity = 1
-        }
-        elseif ($applicationPoolIdentityType -eq 'NetworkService')
-        {
-            $appPoolIdentity = 2
-        }
+        Write-Verbose -Message "Application Pool $appPool already exists"
     }
+    else
+    {
+        Write-Verbose -Message "Adding App Pool [$appPool]"
+        $null = New-WebAppPool -Name $appPool
 
-    $appPoolItem = Get-Item IIS:\AppPools\$appPool
-    $appPoolItem.managedRuntimeVersion = 'v4.0'
-    $appPoolItem.enable32BitAppOnWin64 = $enable32BitAppOnWin64
-    $appPoolItem.processModel.identityType = $appPoolIdentity
-    $appPoolItem | Set-Item
+        Write-Verbose -Message 'Set App Pool Properties'
+        $appPoolIdentity = 4
+        if ($applicationPoolIdentityType)
+        {
+            # LocalSystem = 0, LocalService = 1, NetworkService = 2, SpecificUser = 3, ApplicationPoolIdentity = 4
+            switch ($applicationPoolIdentityType)
+            {
+                'LocalSystem'
+                {
+                    $appPoolIdentity = 0
+                }
+                'LocalService'
+                {
+                    $appPoolIdentity = 1
+                }
+                'NetworkService'
+                {
+                    $appPoolIdentity = 2
+                }
+                'ApplicationPoolIdentity'
+                {
+                    $appPoolIdentity = 4
+                }
+                default {
+                    throw "Invalid value [$applicationPoolIdentityType] for parameter -applicationPoolIdentityType"
+                }
+            }
+        }
+
+        $appPoolItem = Get-Item IIS:\AppPools\$appPool
+        $appPoolItem.managedRuntimeVersion = 'v4.0'
+        $appPoolItem.enable32BitAppOnWin64 = $enable32BitAppOnWin64
+        $appPoolItem.processModel.identityType = $appPoolIdentity
+        $appPoolItem | Set-Item
+    }
 
     Write-Verbose -Message 'Add and Set Site Properties'
     if ($certificateThumbPrint -eq 'AllowUnencryptedTraffic')
@@ -529,6 +617,11 @@ function New-PSWSEndpoint
         [System.String]
         $app = 'PSWS',
 
+        # IIS Application Name for the Site
+        [Parameter()]
+        [System.String]
+        $appPool,
+
         # IIS App Pool Identity Type - must be one of LocalService, LocalSystem, NetworkService, ApplicationPoolIdentity
         [Parameter()]
         [ValidateSet('LocalService', 'LocalSystem', 'NetworkService', 'ApplicationPoolIdentity')]
@@ -603,6 +696,11 @@ function New-PSWSEndpoint
         $Enable32BitAppOnWin64 = $false
     )
 
+    if (-not $appPool)
+    {
+        $appPool = $DefaultAppPoolName
+    }
+
     $script:wevtutil = "$env:windir\system32\Wevtutil.exe"
 
     $svcName = Split-Path $svc -Leaf
@@ -616,12 +714,24 @@ function New-PSWSEndpoint
     $cimInstance = Get-CimInstance -ClassName Win32_ComputerSystem -Verbose:$false
 
     Write-Verbose ("Setting up endpoint at - $protocol//" + $cimInstance.Name + ':' + $port + '/' + $svcName)
-    Initialize-Endpoint -site $site -path $path -cfgfile $cfgfile -port $port -app $app `
-                        -applicationPoolIdentityType $applicationPoolIdentityType -svc $svc -mof $mof `
-                        -dispatch $dispatch -asax $asax -dependentBinaries $dependentBinaries `
-                        -language $language -dependentMUIFiles $dependentMUIFiles -psFiles $psFiles `
-                        -removeSiteFiles $removeSiteFiles -certificateThumbPrint $certificateThumbPrint `
-                        -enable32BitAppOnWin64 $Enable32BitAppOnWin64
+    Initialize-Endpoint `
+        -site $site `
+        -path $path `
+        -cfgfile $cfgfile `
+        -port $port `
+        -app $app `
+        -applicationPoolIdentityType $applicationPoolIdentityType `
+        -svc $svc `
+        -mof $mof `
+        -dispatch $dispatch `
+        -asax $asax `
+        -dependentBinaries $dependentBinaries `
+        -language $language `
+        -dependentMUIFiles $dependentMUIFiles `
+        -psFiles $psFiles `
+        -removeSiteFiles $removeSiteFiles `
+        -certificateThumbPrint $certificateThumbPrint `
+        -enable32BitAppOnWin64 $Enable32BitAppOnWin64
 
     if ($EnablePSWSETW)
     {
@@ -867,4 +977,6 @@ function Set-BindingRedirectSettingInWebConfig
     }
 }
 
-Export-ModuleMember -function New-PSWSEndpoint, Set-AppSettingsInWebconfig, Set-BindingRedirectSettingInWebConfig, Remove-PSWSEndpoint
+Export-ModuleMember `
+    -Function New-PSWSEndpoint, Set-AppSettingsInWebconfig, Set-BindingRedirectSettingInWebConfig, Remove-PSWSEndpoint `
+    -Variable DefaultAppPoolName
